+++ packages/demo/src/components/Settings/Settings.tsx
@@ -14,7 +14,13 @@ const SelectBuilderOption: FC<{
 }> = ({ builder, name }) => <option value={builder}>{name}</option>;
 
 const SelectBuilder: FC = () => {
+<<<<<<< HEAD
   const { setSDK, setBuilderLib, builderLib, useV3Contracts } = useAppState();
+=======
+  const { setSDK, useV3Contracts } = useAppState();
+  const [builderLib, setBuilderLib] = useState<TSupportedTxBuilders>("lucid");
+  const [network, setNetwork] = useState<0 | 1>(0);
+>>>>>>> f79bdb010346a6b2b7fb57a4dbbd70054c785826
 
   const handleTxBuilderLoaderSelect = (key: ETxBuilderType) => {
     setBuilderLib(key);
@@ -60,20 +66,22 @@ const SelectBuilder: FC = () => {
           const { Lucid, Blockfrost } = await import("lucid-cardano");
           const lucidInstance = await Lucid.new(
             new Blockfrost(
-              "https://cardano-preview.blockfrost.io/api/v0/",
+              "https://cardano-mainnet.blockfrost.io/api/v0/",
               // @ts-ignore
               window.__APP_CONFIG.blockfrostAPI
             ),
-            "Preview"
+            network ? "Mainnet" : "Preview"
           );
 
           const options: ISundaeSDKOptions = {
             customQueryProvider: !useV3Contracts
-              ? new QueryProviderSundaeSwapLegacy("preview")
+              ? new QueryProviderSundaeSwapLegacy(
+                  network ? "mainnet" : "preview"
+                )
               : undefined,
             wallet: {
               name: "eternl",
-              network: "preview",
+              network: network ? "mainnet" : "preview",
               builder: {
                 lucid: lucidInstance,
                 type: ETxBuilderType.LUCID,
@@ -88,7 +96,7 @@ const SelectBuilder: FC = () => {
 
       setSDK(sdk);
     })();
-  }, [builderLib, setSDK, useV3Contracts]);
+  }, [builderLib, setSDK, useV3Contracts, network]);
 
   return (
     <div className="container flex gap-10">
@@ -106,6 +114,17 @@ const SelectBuilder: FC = () => {
           <option value="undefined">None</option>
           <SelectBuilderOption builder={ETxBuilderType.LUCID} name="Lucid" />
           <SelectBuilderOption builder={ETxBuilderType.BLAZE} name="Blaze" />
+        </select>
+      </div>
+      <div className="p-4">
+        <h2 className="mb-4 text-lg font-bold text-white">Network</h2>
+        <select
+          className="mr-4 w-full rounded-md bg-slate-800 px-4 py-2"
+          value={network}
+          onChange={(e) => setNetwork(Number(e.target.value) as 0 | 1)}
+        >
+          <option value="0">Preview</option>
+          <option value="1">Mainnet</option>
         </select>
       </div>
     </div>